Ash/expected

+++ fun_test/web/fun_test/metrics_models.py
@@ -938,7 +938,7 @@ class NuTransitPerformance(models.Model):
     output_jitter_min = models.FloatField(verbose_name="Jitter min in us", default=0)
     output_jitter_max = models.FloatField(verbose_name="Jitter max in us", default=0)
     output_jitter_avg = models.FloatField(verbose_name="Jitter avg in us", default=0)
-    output_pps = models.IntegerField(verbose_name="Packets per sec", default=0)
+    output_pps = models.FloatField(verbose_name="Packets per sec", default=0)
     input_mode = models.CharField(verbose_name="Port modes (25, 50 or 100 G)", max_length=20, default="")
     input_version = models.CharField(verbose_name="Version", max_length=50)
     input_flow_type = models.CharField(verbose_name="Flow Type", max_length=50, default="")
